upd: collectibles

+++ Assets/_Source/BallSystem/BallScorer.cs
@@ -5,9 +5,9 @@ namespace BallSystem
     public class BallScorer
     {
         public event Action OnBallScoreChange;
-        public int Score { get; private set; }
+        public float Score { get; private set; }
 
-        public void AddScore(int value)
+        public void AddScore(float value)
         {
             Score += value;
             OnBallScoreChange?.Invoke();
